// hive/github-universe/mirror/src/client/src/lib/pulselang/ast.ts
export interface ProgramNode {
  type: "Program";
  universeOp: string;
  contextId: string;
  statements: StatementNode[];
}

export type StatementNode =
  | FieldDeclNode
  | AssignNode
  | ReturnNode
  | FusionNode;

export interface FieldDeclNode {
  type: "FieldDecl";
  name: string;
  typeName: string;
}

export interface AssignNode {
  type: "Assign";
  target: string;
  callExpr: CallExprNode;
}

export interface ReturnNode {
  type: "Return";
  name: string;
}

export interface FusionNode {
  type: "Fusion";
  left: string;
  right: string;
  target: string;
}

export interface CallExprNode {
  type: "CallExpr";
  callee: string;
  args: ArgNode[];
}

export interface ArgNode {
  type: "Arg";
  key: string;
  value: string | CallExprNode;
}